+++ tests/based_on_issue/mod.rs
@@ -1,0 +1,1 @@
+pub mod fediverse_handle_82;